+++ src/gui/group_component.cpp
@@ -45,11 +45,12 @@ GroupComponent::draw()
   if (!title.empty())
     {
       TTFFont* font = Fonts::current()->vera20.get();
-      font->draw_center(Vector2f(rect.left + rect.get_width()/2, rect.top + font->get_height() + 5), 
+      font->draw_center(Vector2f(rect.left + static_cast<float>(rect.get_width()) / 2.0f,
+                                 rect.top  + static_cast<float>(font->get_height()) + 5.0f),
                         title, Color(1.0f, 1.0f, 1.0f));
 
-      Display::fill_rect(Rectf(rect.left  + 8, rect.top + font->get_height() + 16,
-                               rect.right - 8, rect.top + font->get_height() + 18),
+      Display::fill_rect(Rectf(rect.left  + 8.0f, rect.top + static_cast<float>(font->get_height()) + 16.0f,
+                               rect.right - 8.0f, rect.top + static_cast<float>(font->get_height()) + 18.0f),
                          Color(1.0f, 1.0f, 1.0f, 0.5f));
     }
 
@@ -77,10 +78,10 @@ GroupComponent::pack(Component* component)
 Rectf
 GroupComponent::get_child_rect() const
 {
-  int padding = 6;
+  float padding = 6.0f;
 
   return Rectf(rect.left   + padding,
-               rect.top    + padding + (title.empty() ? 0 : Fonts::current()->vera20->get_height() + 18),
+               rect.top    + padding + (title.empty() ? 0.0f : static_cast<float>(Fonts::current()->vera20->get_height()) + 18.0f),
                rect.right  - padding,
                rect.bottom - padding);
 }
