+++ src/gui/menu.cpp
@@ -103,7 +103,7 @@ Menu::show()
                     (rect.top + rect.bottom) / 2.0f);
 
     Sizef size(menu->get_prefered_width(), 
-               menu->get_prefered_height() + (group->has_title() ? Fonts::current()->vera20->get_height() + 18 : 0.0f));
+               menu->get_prefered_height() + (group->has_title() ? static_cast<float>(Fonts::current()->vera20->get_height()) + 18.0f : 0.0f));
 
     group->set_screen_rect(Rectf(Vector2f(center.x - size.width/2.0f,
                                           center.y - size.height/2.0f),
