+++ src/gui/tab_component.cpp
@@ -48,12 +48,12 @@ TabComponent::draw()
 {
   if (tabs.empty()) return;
 
-  float tab_width = rect.get_width()/tabs.size();
+  float tab_width = rect.get_width() / static_cast<float>(tabs.size());
   for(int i = 0; i != int(tabs.size()); ++i)
     {
-      Rectf tab_rect(Vector2f(rect.left + tab_width * i + 10,
-                            rect.top),
-                     Sizef(tab_width - 20, Fonts::current()->vera20->get_height() + 6));
+      Rectf tab_rect(Vector2f(rect.left + tab_width * static_cast<float>(i) + 10.0f,
+                              rect.top),
+                     Sizef(tab_width - 20.0f, static_cast<float>(Fonts::current()->vera20->get_height()) + 6.0f));
 
       if (i == current_tab)
         Display::fill_rounded_rect(tab_rect, 5.0f, Color(1.0f, 1.0f, 1.0f, 0.5f));
@@ -62,8 +62,8 @@ TabComponent::draw()
 
       Display::draw_rounded_rect(tab_rect, 5.0f, Color(1.0f, 1.0f, 1.0f, 0.5f));
 
-      Fonts::current()->vera20->draw_center(Vector2f(rect.left + tab_width * i + tab_width/2,
-                                                     rect.top + Fonts::current()->vera20->get_height()),
+      Fonts::current()->vera20->draw_center(Vector2f(rect.left + tab_width * static_cast<float>(i) + tab_width/2,
+                                                     rect.top + static_cast<float>(Fonts::current()->vera20->get_height())),
                                             tabs[i].label,
                                             tabs[current_tab].component->is_active()
                                             ? Color(1.0f, 1.0f, 1.0f, 0.5f) 
@@ -139,9 +139,9 @@ TabComponent::pack(const std::string& name, Component* component)
 {
   tabs.push_back(Tab(name, component));
 
-  int padding = 6;
+  float padding = 6.0f;
   component->set_screen_rect(Rectf(rect.left + padding,
-                                   rect.top  + padding + Fonts::current()->vera20->get_height() + 10.0f,
+                                   rect.top  + padding + static_cast<float>(Fonts::current()->vera20->get_height()) + 10.0f,
                                    rect.right  - padding,
                                    rect.bottom - padding
                                    ));
